+++ ihp-hsx/default.nix
@@ -1,4 +1,5 @@
 { mkDerivation, base, blaze-html, blaze-markup, bytestring
+, bytestring-strict-builder
 , containers, deepseq, ghc, hspec, lib, lucid2, megaparsec, mtl
 , process, string-conversions, tasty-bench, template-haskell, text
 , transformers, unordered-containers
@@ -10,7 +11,8 @@ mkDerivation {
   isLibrary = true;
   isExecutable = true;
   libraryHaskellDepends = [
-    base blaze-html blaze-markup bytestring containers ghc lucid2
+    base blaze-html blaze-markup bytestring bytestring-strict-builder
+    containers ghc lucid2
     megaparsec string-conversions template-haskell text transformers
     unordered-containers
   ];
@@ -19,11 +21,13 @@ mkDerivation {
     string-conversions template-haskell text unordered-containers
   ];
   testHaskellDepends = [
-    base blaze-markup bytestring containers hspec lucid2 megaparsec mtl
-    process string-conversions template-haskell text unordered-containers
+    base blaze-markup bytestring bytestring-strict-builder containers
+    hspec lucid2 megaparsec mtl process
+    string-conversions template-haskell text unordered-containers
   ];
   benchmarkHaskellDepends = [
-    base bytestring containers deepseq megaparsec string-conversions
+    base blaze-html blaze-markup bytestring bytestring-strict-builder
+    containers deepseq megaparsec string-conversions
     tasty-bench template-haskell text unordered-containers
   ];
   doHaddock = false;